+++ app/src/main/java/org/ole/planet/myplanet/ui/team/teamMember/JoinedMemberFragment.java
@@ -12,6 +12,7 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import org.ole.planet.myplanet.R;
+import org.ole.planet.myplanet.base.BaseMemberFragment;
 import org.ole.planet.myplanet.model.RealmMyTeam;
 import org.ole.planet.myplanet.model.RealmUserModel;
 import org.ole.planet.myplanet.ui.team.BaseTeamFragment;
@@ -21,29 +22,26 @@ import java.util.List;
 /**
  * A simple {@link Fragment} subclass.
  */
-public class JoinedMemberFragment extends BaseTeamFragment {
+public class JoinedMemberFragment extends BaseMemberFragment {
 
-    RecyclerView rvMember;
-    TextView tvNodata;
+
     public JoinedMemberFragment() {
     }
 
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        View v =  inflater.inflate(R.layout.fragment_members, container, false);
-        rvMember = v.findViewById(R.id.rv_member);
-        tvNodata = v.findViewById(R.id.tv_nodata);
-        return v;
+    public List<RealmUserModel> getList() {
+        return RealmMyTeam.getJoinedMemeber(teamId, mRealm);
     }
 
     @Override
-    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-        List<RealmUserModel> joinedMember = RealmMyTeam.getJoinedMemeber(teamId, mRealm);
-        rvMember.setLayoutManager(new GridLayoutManager(getActivity(), 3));
-        rvMember.setAdapter(new AdapterJoinedMemeber(getActivity(),joinedMember, mRealm));
-        showNoData(tvNodata, joinedMember.size());
+    public RecyclerView.Adapter getAdapter() {
+        return new AdapterJoinedMemeber(getActivity(),getList(), mRealm);
     }
+
+    @Override
+    public RecyclerView.LayoutManager getLayoutManager() {
+        return new GridLayoutManager(getActivity(), 3);
+    }
+
 }
